refactor to separate roomprompt and remove extra console logs
