+++ socket/client/authPrompt.js
@@ -66,12 +66,12 @@ module.exports = (socket, SERVER) => async function authPrompt() {
 
   //signin logic
   const authAnswers = await inquirer.prompt(auth[authRoute]);
-  console.log('🚀 ~ file: authPrompt.js:68 ~ authPrompt ~ authAnswers', authAnswers);
+  // console.log('🚀 ~ file: authPrompt.js:68 ~ authPrompt ~ authAnswers', authAnswers);
   const { username, password } = authAnswers;
 
   if(authRoute === 'signin') {
     const encoded = base64.encode(`${username}:${password}`);
-    console.log('🚀 ~ file: authPrompt.js:65 ~ authPrompt ~ encoded', encoded);
+    // console.log('🚀 ~ file: authPrompt.js:65 ~ authPrompt ~ encoded', encoded);
     const response = await axios.post(`${SERVER}/signin`,{}, {
       headers: {
         Authorization: `Basic ${encoded}`,
@@ -94,7 +94,7 @@ module.exports = (socket, SERVER) => async function authPrompt() {
       }
     });
 
-    console.log(body);
+    // console.log(body);
     try {
       const response = await axios.post(`${SERVER}/signup`, body);
       return response.data;
